Moved and renamed BottomSheetPlayer code to MiniPlayer
