+++ app/src/main/java/com/example/music/ui/albumdetails/AlbumDetailsScreen.kt
@@ -84,7 +84,7 @@ import com.example.music.domain.model.SongInfo
 import com.example.music.domain.testing.PreviewSongs
 import com.example.music.ui.player.MiniPlayerControlActions
 import com.example.music.ui.shared.AlbumMoreOptionsBottomModal
-import com.example.music.ui.shared.BottomSheetPlayer
+import com.example.music.ui.shared.MiniPlayer
 import com.example.music.ui.shared.DetailsSortSelectionBottomModal
 import com.example.music.ui.shared.Error
 import com.example.music.ui.shared.ItemCountAndSortSelectButtons
@@ -301,7 +301,7 @@ fun AlbumDetailsScreen(
             },
             bottomBar = {
                 if (isActive){
-                    BottomSheetPlayer(
+                    MiniPlayer(
                         song = currentSong,
                         isPlaying = isPlaying,
                         navigateToPlayer = navigateToPlayer,
